Web:Store:HotkeyStore Fixed cyclic dependencies
