+++ packages/client/src/store/HotkeyStore.js
@@ -1,13 +1,16 @@
-import { combineUrl } from "@docspace/common/utils";
-import { isDesktop, isMobile } from "@docspace/components/utils/device";
 import { makeAutoObservable } from "mobx";
+
+import { combineUrl } from "@docspace/common/utils";
+import { RoomsType } from "@docspace/common/constants";
+import { checkDialogsOpen } from "@docspace/common/utils/checkDialogsOpen";
+
+import toastr from "@docspace/components/toast/toastr";
+import { isDesktop, isMobile } from "@docspace/components/utils/device";
+import getFilesFromEvent from "@docspace/components/drag-and-drop/get-files-from-event";
+
 import config from "PACKAGE_FILE";
 import { getCategoryUrl } from "SRC_DIR/helpers/utils";
-import toastr from "@docspace/components/toast/toastr";
-import { RoomsType } from "@docspace/common/constants";
-import { encryptionUploadDialog } from "../helpers/desktop";
-import getFilesFromEvent from "@docspace/components/drag-and-drop/get-files-from-event";
-import { checkDialogsOpen } from "@docspace/common/utils/checkDialogsOpen";
+import { encryptionUploadDialog } from "../helpers/encryptionUploadDialog";
 
 class HotkeyStore {
   filesStore;
@@ -533,11 +536,14 @@ class HotkeyStore {
       folderInput && folderInput.click();
     } else {
       if (this.treeFoldersStore.isPrivacyFolder) {
-        encryptionUploadDialog((encryptedFile, encrypted) => {
-          encryptedFile.encrypted = encrypted;
-          this.goToHomePage(navigate);
-          this.uploadDataStore.startUpload([encryptedFile], null, t);
-        });
+        encryptionUploadDialog(
+          this.settingsStore.extsWebEncrypt,
+          (encryptedFile, encrypted) => {
+            encryptedFile.encrypted = encrypted;
+            this.goToHomePage(navigate);
+            this.uploadDataStore.startUpload([encryptedFile], null, t);
+          }
+        );
       } else {
         const fileInput = document.getElementById("customFileInput");
         fileInput && fileInput.click();
